+++ ironbank/build_ironbank_docker.sh
@@ -63,14 +63,14 @@ function build_hardening_manifest {
     if [[ ! -d $OUTPUT_PATH ]]; then
       mkdir $OUTPUT_PATH
     fi
-    BASE_IMAGE=`python ./ironbank/get_docker_image_python_version.py --docker_image_dir $1`
-    TAG="3.9.6.22912"
-    PYTHON_VERSION="3"
-    if [[ "$BASE_IMAGE" == "python" ]]; then
-      TAG="2.7.18.20958"
-      PYTHON_VERSION="2"
+    BASE_IMAGE_NAME_AND_TAG=`python ./ironbank/get_ironbank_base_image_name_and_tag.py --docker_image_dir $1`
+    echo "Base image name and tag: $BASE_IMAGE_NAME_AND_TAG"
+    PYTHON_VERSION="2"
+    if [[ -n $(echo $BASE_IMAGE_NAME_AND_TAG | grep 'python3') ]]; then
+      PYTHON_VERSION="3"
     fi
-    DOCKER_IMAGE="$REGISTRYONE_URL/ironbank/opensource/palo-alto-networks/demisto/$BASE_IMAGE:$TAG"
+    DOCKER_IMAGE="$REGISTRYONE_URL/$BASE_IMAGE_NAME_AND_TAG"
+    echo "Docker image is $DOCKER_IMAGE"
     docker pull $DOCKER_IMAGE
     DOCKER_PACKAGES_METADATA_PATH="$OUTPUT_PATH/docker_packages_metadata.txt"
     REQUIREMENTS="$(cat $1/requirements.txt)"
@@ -146,6 +146,9 @@ function commit_ironbank_image_to_repo_one {
   fi
   cp -r $CURRENT_DIR/ironbank/$IMAGE_NAME/* .
   cp $CURRENT_DIR/docker/$IMAGE_NAME/requirements.txt .
+  pwd
+  ls
+  sed -i -e '/^-i/d' requirements.txt # remove -i from requirements file
   git config user.email "containers@demisto.com"
   git config user.name "dc-builder"
   if [[ $(git diff --exit-code) ]]; then
